Extract the fancy wall clock `now()` into its own class.
