+++ src/clocks/README.md
@@ -1,0 +1,12 @@
+@this/clocks
+============
+
+Interface to real-world clocks. This is as opposed to simply the mechanics of
+time/date calculation, which in this project resides in the `data-values`
+module.
+
+- - - - - - - - - -
+```
+Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
+SPDX-License-Identifier: Apache-2.0
+```
