+++ src/clocks/index.js
@@ -1,0 +1,4 @@
+// Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
+// SPDX-License-Identifier: Apache-2.0
+
+export * from '#x/WallClock';
