+++ src/clocks/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/loggy",
+  "name": "@this/clocks",
   "version": "0.0.1",
   "type": "module",
   "private": true,
@@ -12,10 +12,6 @@
   },
 
   "dependencies": {
-    "@this/async": "*",
-    "@this/clocks": "*",
-    "@this/data-values": "*",
-    "@this/metacomp": "*",
-    "@this/typey": "*"
+    "@this/data-values": "*"
   }
 }
